// helm-workspace/apps/api-shared-helm/src/models/solucionAmbitoSector.ts
export interface SolucionAmbitoSector 
{
    id_ambito?: number;
    id_solucion?: number;
    id_sector?: number;
    description: string;
    title: string;
    subtitle: string;
    icon: string;
    titleweb: string;
    slug: string;
    multimediaUri: string;
    multimediaTypeId: string;
    problemaTitle: string;
    problemaPragma: string;
    solucionTitle: string;
    solucionPragma: string;
    caracteristicasTitle: string;
    caracteristicasPragma: string;
    casosdeusoTitle: string;
    casosdeusoPragma: string;
    firstCtaTitle: string;
    firstCtaPragma: string;
    secondCtaTitle: string;
    secondCtaPragma: string;
    beneficiosTitle: string;
    beneficiosPragma: string;
}